refactor: 拆分用户详情页和集换交易面板为子组件
- pages/user/[wikidotId].vue (1441→833行): 提取 UserHeader, UserWorks,
  UserPreferences, UserVotes, UserRevisions 5个子组件
- GachaTradePanel.vue (2637→1094行): 提取 TradeSellForm, TradeListings,
  TradeBuyForm, TradeDetail 4个子组件
- 数据获取(useAsyncData)保留在页面/父组件级别，子组件通过 props 接收数据
- 交易面板CSS从 scoped 改为 .trade-bazaar 命名空间以允许子组件继承样式
- 构建验证通过，无业务逻辑变更
